Funnel animation: Insulin, Carbs, Bolus Wizard, TempTarget
