+++ ui/src/main/res/anim/insulin_dialog_funnel_up.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate
+        android:fromYDelta="100%"
+        android:toYDelta="0%"
+        android:duration="@android:integer/config_shortAnimTime"
+        android:interpolator="@android:anim/accelerate_interpolator"/>
+    <scale
+        android:fromXScale="0.01"
+        android:fromYScale="0.01"
+        android:toXScale="1.0"
+        android:toYScale="1.0"
+        android:duration="@android:integer/config_shortAnimTime"
+        android:pivotX="12%"
+        android:pivotY="190%"/>
+    <alpha
+        android:fromAlpha="0.1"
+        android:toAlpha="1.0"
+        android:duration="@android:integer/config_shortAnimTime"
+        android:interpolator="@android:anim/accelerate_decelerate_interpolator"/>
+</set>
