+++ ui/src/main/res/anim/wiz_dialog_funnel_down.xml
@@ -1,21 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
 <set xmlns:android="http://schemas.android.com/apk/res/android">
     <translate
-        android:fromYDelta="100%"
-        android:toYDelta="0%"
+        android:fromYDelta="0%"
+        android:toYDelta="100%"
         android:duration="@android:integer/config_shortAnimTime"
-        android:interpolator="@android:anim/accelerate_interpolator"/>
+        android:interpolator="@android:anim/decelerate_interpolator"/>
     <scale
-        android:fromXScale="0.01"
-        android:fromYScale="0.01"
-        android:toXScale="1.0"
-        android:toYScale="1.0"
+        android:fromXScale="1.0"
+        android:fromYScale="1.0"
+        android:toXScale="0.01"
+        android:toYScale="0.01"
         android:duration="@android:integer/config_shortAnimTime"
-        android:pivotX="60%"
-        android:pivotY="180%"/>
+        android:pivotX="110%"
+        android:pivotY="150%"/>
     <alpha
-        android:fromAlpha="0.1"
-        android:toAlpha="1.0"
+        android:fromAlpha="1.0"
+        android:toAlpha="0.1"
         android:duration="@android:integer/config_shortAnimTime"
         android:interpolator="@android:anim/accelerate_decelerate_interpolator"/>
 </set>
